requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic[email]
pydantic-settings
python-jose[cryptography]
bcrypt  
python-multipart
python-dotenv
alembic
requests
httpx
aiofiles
stripe